requirements.txt
asgiref==3.3.1
async-timeout==4.0.2
Django==3.0.7
django-cors-headers==4.0.0
django-redis==3.1.6
djangorestframework==3.12.2
Faker==4.14.2
Pillow==9.5.0
python-dateutil==2.8.1
python-dotenv==1.0.0
pytz==2020.4
redis==4.5.5
six==1.15.0
sqlparse==0.4.1
text-unidecode==1.3
typing_extensions==4.6.2
validate-docbr==1.8.2
